[FLINK-2890] [tests] Apply JMH on StringSerializationSpeedBenchmark class.
